perf(computeRanges): compute for all components simultaneously
