+++ src/IO/Analyze/UpdateHistogram.worker.js
