+++ src/IO/Analyze/webWorkerPromiseWorkerPool.js
