+++ src/IO/InMemoryMultiscaleSpatialImage.js
@@ -8,8 +8,8 @@ import {
   imageSharedBufferOrCopy,
   stackImages,
 } from 'itk-wasm'
-import computeRange from '../Rendering/VTKJS/computeRange'
 import { chunkArray, CXYZT, orderBy, toDimensionMap } from './dimensionUtils'
+import { computeRanges } from './Analyze/computeRanges'
 
 const numberOfWorkers = navigator.hardwareConcurrency
   ? navigator.hardwareConcurrency
@@ -190,11 +190,9 @@ async function chunkImage(image, chunkSize) {
     //}
   }
 
-  const ranges = []
-  for (let comp = 0; comp < sizeCXYZTElements[0]; comp++) {
-    const range = await computeRange(image.data, comp, sizeCXYZTElements[0])
-    ranges.push([range.min, range.max])
-  }
+  const ranges = (
+    await computeRanges(image.data, sizeCXYZTElements[0])
+  ).map(range => [range.min, range.max])
 
   const orderByDims = orderBy(dims)
   const scaleInfo = {
